% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeGraphsFunction.R
\name{.get_clean_gene_annotations}
\alias{.get_clean_gene_annotations}
\title{Internal: load and clean GENCODE gene annotations}
\usage{
.get_clean_gene_annotations(genome = "hg38")
}
\arguments{
\item{genome}{"hg19" or "hg38"}
}
\description{
Returns a GRanges with one range per gene, including metadata:
  - gene      : gene symbol (gencode gene_name)
  - hgnc_id   : HGNC identifier (if available, otherwise NA)
  - mane_type : MANE type (e.g., "MANE Select", "MANE Plus Clinical", or NA)
}
\keyword{internal}
