% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeGraphsFunction.R
\name{get_common_genes_SV}
\alias{get_common_genes_SV}
\title{Get genes under SVs common across VCFs}
\usage{
get_common_genes_SV(vcf_files, sv_type = "DEL", genome = "hg38")
}
\arguments{
\item{vcf_files}{A character vector of VCF file paths.}

\item{sv_type}{Type of SV (e.g., "DEL").}

\item{genome}{Genome assembly ("hg19" or "hg38").}
}
\description{
Returns genes impacted by SVs shared among multiple VCF files.
}
\examples{
# files <- GeGraphs:::load_example_files_HG02317()
# get_common_genes_SV(
#   vcf_files = c(files$delly, files$manta, files$del),
#   sv_type   = "DEL",
#   genome    = "hg38"
# )
}
