% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GeGraphsFunction.R
\name{get_genes_SV}
\alias{get_genes_SV}
\title{Get genes overlapped by SVs}
\usage{
get_genes_SV(vcf_file, sv_type = "DEL", genome = "hg38")
}
\arguments{
\item{vcf_file}{Path to VCF file.}

\item{sv_type}{Type of SV to consider (e.g., "DEL").}

\item{genome}{Genome assembly ("hg19" or "hg38").}
}
\description{
Returns gene names that are overlapped by structural variants in a VCF file.
}
\examples{
# files <- GeGraphs:::load_example_files_HG02317()
# get_genes_SV(files$delly, sv_type = "DEL", genome = "hg38")
}
